feat: add node-twitter unfollow --after-error
A user encountered the following error:
```
Unfollowing 66 of 225: screen_name (#144)...
RethrownError: connect ENETDOWN 104.244.42.2:443 - Local (192.168.1.160:62601)
    at processCommand (/snapshot/tmp/lib/NodeTwitterMain.js:60:11)
    at processTicksAndRejections (internal/process/task_queues.js:97:5)
RethrownError: connect ENETDOWN 104.244.42.2:443 - Local (192.168.1.160:62601)
    at NodeTwitter.unfollow (/snapshot/tmp/lib/NodeTwitter.js:153:13)
    at processTicksAndRejections (internal/process/task_queues.js:97:5)
    at async processCommand (/snapshot/tmp/lib/NodeTwitterMain.js:45:9)
RethrownError: connect ENETDOWN 104.244.42.2:443 - Local (192.168.1.160:62601)
    at /snapshot/tmp/lib/NodeTwitter.js:612:11
    at processTicksAndRejections (internal/process/task_queues.js:97:5)
    at async NodeTwitter.unfollow (/snapshot/tmp/lib/NodeTwitter.js:122:7)
    at async processCommand (/snapshot/tmp/lib/NodeTwitterMain.js:45:9)
Error: connect ENETDOWN 104.244.42.2:443 - Local (192.168.1.160:62601)
    at internalConnect (net.js:919:16)
    at defaultTriggerAsyncIdScope (internal/async_hooks.js:378:12)
    at GetAddrInfoReqWrap.emitLookup [as callback] (net.js:1059:9)
    at GetAddrInfoReqWrap.onlookup [as oncomplete] (dns.js:68:8)
```

For users with large Twitter followings, it may take a long time
to create the friends list that we use to determine whom to
unfollow. If node-twitter created a list of users to unfollow and
an error occurs while unfollowing users on that list, call
`node-twitter unfollow --after-error` to unfollow the latest list of
users without having to wait to determine whom to unfollow.